Can now specify what text to clear in purge command
Purge command now takes an optional "content_to_delete" argument, which if not None, will search through `limit` messages but only messages that contain the given phrase will be deleted.

+++ cogs/moderation.py
@@ -15,21 +15,20 @@ class Moderation(commands.Cog, name="Mod Only"):
         description="Takes an argument `limit` and deletes that many messages from the current channel."
     )
     @commands.has_permissions(manage_messages=True)
-    async def purge(self, ctx, limit: int):
+    async def purge(self, ctx, limit: int, *, content_to_delete=None):
         await ctx.message.delete()
-        
         message_count = {}
 
         async for message in ctx.channel.history(limit=limit):
-            if message.author.name in message_count.keys():
-                message_count[message.author.name] += 1
-            else:
-                message_count[message.author.name] = 1
-
-            await message.delete()
+            if content_to_delete is None or content_to_delete.lower() in message.content.lower():
+                if message.author in message_count.keys():
+                    message_count[message.author] += 1
+                else:
+                    message_count[message.author] = 1
+                await message.delete()
         
         deleted = sum(message_count.values())
-        messages = [f"{deleted} message{' was' if deleted == 1 else 's were'} removed."]
+        messages = [f"{f'*Only messages containing the phrase `{content_to_delete}` were deleted*' if content_to_delete is not None else ''}\n{deleted} message{' was' if deleted == 1 else 's were'} removed."]
         messages.extend(f"- **{author}**: {count}" for author, count in message_count.items())
         await ctx.send("\n".join(messages), delete_after=10)
